Code
The module shown in the image is an I2C interface for an LCD display, commonly used to simplify connections to an Arduino and reduce the number of required pins. This type of module is usually used to control an LCD display (like a 16x2 or 20x4) in Arduino projects, such as your RFID and LED project.